refac: move task to environment
