+++ api/tests/unit/audit/test_unit_audit_models.py
@@ -160,67 +160,33 @@ def test_audit_log_save_project_is_added_if_not_set(environment):
     assert audit_log.project == environment.project
 
 
-def test_creating_audit_logs_triggers_environment_update_message(environment, mocker):
+def test_creating_audit_logs_creates_process_environment_update_task(
+    environment, mocker
+):
     # Given
-    send_environment_update_message_for_environment = mocker.patch(
-        "audit.models.send_environment_update_message_for_environment"
+    process_environment_update = mocker.patch(
+        "environments.tasks.process_environment_update"
     )
 
     # When
     audit_log = AuditLog.objects.create(environment=environment)
 
     # Then
-    send_environment_update_message_for_environment.assert_called_once_with(environment)
+    process_environment_update.delay.assert_called_once_with(args=(audit_log.id,))
+
+    # and
+    environment.refresh_from_db()
     assert environment.updated_at == audit_log.created_date
 
 
-def test_creating_audit_logs_triggers_environment_update_message_for_environment_if_environment_is_present(
+def test_creating_audit_logs_for_change_request_does_not_trigger_process_environment_update(
     environment, mocker, project
 ):
     # Given
-    send_environment_update_message_for_environment = mocker.patch(
-        "audit.models.send_environment_update_message_for_environment"
+    process_environment_update = mocker.patch(
+        "environments.tasks.process_environment_update"
     )
-    send_environment_update_message_for_project = mocker.patch(
-        "audit.models.send_environment_update_message_for_project"
-    )
-    # When
-    audit_log = AuditLog.objects.create(environment=environment, project=project)
 
-    # Then
-    send_environment_update_message_for_environment.assert_called_once_with(environment)
-    send_environment_update_message_for_project.assert_not_called()
-    assert environment.updated_at == audit_log.created_date
-
-
-def test_creating_audit_logs_triggers_environment_update_message_for_project_if_environment_is_not_present(
-    environment, mocker, project
-):
-    # Given
-    send_environment_update_message_for_environment = mocker.patch(
-        "audit.models.send_environment_update_message_for_environment"
-    )
-    send_environment_update_message_for_project = mocker.patch(
-        "audit.models.send_environment_update_message_for_project"
-    )
-    # When
-    AuditLog.objects.create(project=project)
-
-    # Then
-    send_environment_update_message_for_project.assert_called_once_with(project)
-    send_environment_update_message_for_environment.assert_not_called()
-
-
-def test_creating_audit_logs_for_change_request_does_not_trigger_environment_update_message(
-    environment, mocker, project
-):
-    # Given
-    send_environment_update_message_for_environment = mocker.patch(
-        "audit.models.send_environment_update_message_for_environment"
-    )
-    send_environment_update_message_for_project = mocker.patch(
-        "audit.models.send_environment_update_message_for_project"
-    )
     # When
     audit_log = AuditLog.objects.create(
         project=project,
@@ -228,6 +194,5 @@ def test_creating_audit_logs_for_change_request_does_not_trigger_environment_upd
     )
 
     # Then
-    send_environment_update_message_for_environment.assert_not_called()
-    send_environment_update_message_for_project.assert_not_called()
+    process_environment_update.delay.assert_not_called()
     assert audit_log.created_date != environment.updated_at
